src/backend/gateway/pyproject.toml
[project]
name = "gateway"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "aiokafka[lz4]==0.13.0",
    "grpcio==1.76.0",
    "grpcio-tools==1.76.0",
    "litestar[standard,prometheus]==2.20.0",
]

[dependency-groups]
test = [
    "coverage==7.13.3",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
]
